% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTA2DNAString.R
\name{FASTA2DNAstring}
\alias{FASTA2DNAstring}
\title{FASTA2DNAstring}
\usage{
FASTA2DNAstring(file)
}
\arguments{
\item{file}{FASTA file}
}
\value{
DNAString object
}
\description{
Makes an DNAString object from an FASTA file
}
\examples{
dir <-system.file("extdata", package="metavirus")
seq <- FASTA2DNAstring(file.path(dir, "sequences.fasta"))

}
