% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFASTAbyBase.R
\name{filterFASTAbyBase}
\alias{filterFASTAbyBase}
\title{filterFASTAbyBase}
\usage{
filterFASTAbyBase(seq, base, prob, cutoff)
}
\arguments{
\item{seq}{DNAString object from BioString}

\item{base}{Base or Bases used to filter (eg. G or GC or GCC)}

\item{prob}{Frequency percentage (TRUE) or exact count (FALSE)}

\item{cutoff}{Maximum frequency allowed. All sequences below this threshold will be removed.}
}
\value{
DNAString
}
\description{
Subset a FASTA file with multiple sequences based on the frequency of an specific base.
}
\examples{
dir <-system.file("extdata", package="metavirus")
seq <- FASTA2DNAstring(file.path(dir, "sequences.fasta"))
subset <- filterFASTAbyBase(seq, "A", TRUE, 0.3)

}
