% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFASTAbyIUPAC.R
\name{filterFASTAbyIUPAC}
\alias{filterFASTAbyIUPAC}
\title{filterFASTAbyIUPAC}
\usage{
filterFASTAbyIUPAC(seq, prob, cutoff, replace)
}
\arguments{
\item{seq}{DNAString object}

\item{prob}{Frequency percentage (TRUE) or exact count (FALSE)}

\item{cutoff}{Maximum frequency allowed. All sequences below this threshold will be removed.}

\item{replace}{Replace remaining IUPAC bases with N (TRUE or FALSE)}
}
\value{
DNAString
}
\description{
Subset an FASTA file based on IUPAC bases percentages (eg. other bases rather than A,C,G, or T)
}
\examples{
dir <-system.file("extdata", package="metavirus")
seq <- FASTA2DNAstring(file.path(dir, "sequences.fasta"))
subset <- filterFASTAbyIUPAC(seq, TRUE, 0.3, FALSE)

}
