% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFASTAbyLength.R
\name{filterFASTAbyLength}
\alias{filterFASTAbyLength}
\title{filterFASTAbyLength}
\usage{
filterFASTAbyLength(seq, min, max)
}
\arguments{
\item{seq}{DNAString object}

\item{min}{Minimum length}

\item{max}{Maximum length}
}
\value{
DNAString object
}
\description{
Subset sequences based on length
}
\examples{
dir <-system.file("extdata", package="metavirus")
seq <- FASTA2DNAstring(file.path(dir, "sequences.fasta"))
subset <- filterFASTAbyLength(seq, 0, 1200)

}
