% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQsStatsReport.R
\name{SEQsStatsReport}
\alias{SEQsStatsReport}
\title{SEQsStatsReport}
\usage{
SEQsStatsReport(seq)
}
\arguments{
\item{seq}{DNAString object}
}
\value{
dataframe
}
\description{
Compile basic info about sequences
}
\examples{
dir <-system.file("extdata", package="metavirus")
seq <- FASTA2DNAstring(file.path(dir, "sequences.fasta"))
subset <- SEQsStatsReport(seq)

}
